requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
pydantic==2.5.2
python-dotenv==1.0.0
aiofiles==23.2.1
alembic==1.12.1
pytest==7.4.3
httpx==0.25.2 